+++ lib/src/processors/ios/xcconfig/ios_xcconfig_processor.dart
@@ -74,6 +74,7 @@ class IOSXCConfigProcessor extends StringProcessor {
       'ASSET_PREFIX': Variable(value: _flavorName),
       'BUNDLE_NAME': Variable(value: _flavor.app.name),
       'BUNDLE_DISPLAY_NAME': Variable(value: _flavor.app.name),
+      'PRODUCT_BUNDLE_IDENTIFIER': Variable(value: _flavor.ios!.bundleId),
     })
       ..addAll(
         _flavor.ios?.variables.where((_, variable) =>